refactor: migrated /api/datasets/[id]/style/[layer]/[type]
