+++ sites/geohub/static/api/swagger/spec.yaml
@@ -529,151 +529,6 @@ paths:
           in: query
           name: compress
           description: 'optional. default is false. If true, compress response as gzip using compress api. Only available for either json or geojson format.'
-  '/datasets/{id}/style/{layer}/{type}':
-    parameters:
-      - $ref: '#/components/parameters/dataset_id'
-      - $ref: '#/components/parameters/maplibre_layer_type'
-      - $ref: '#/components/parameters/maplibre_layer_id'
-    get:
-      summary: Get default layer style for a dataset
-      tags:
-        - datasets
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/DatasetLayerStyle'
-        '404':
-          description: Not Found
-        '':
-          description: ''
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  dataset_id:
-                    type: string
-                  layer_type:
-                    type: string
-                  source:
-                    type: string
-                    description: JSON object for maplibre source
-      operationId: get-datasets-id-style
-      description: |-
-        This endpoint is to return the default layer style for specified dataset if the style infomation is registered in the database.
-
-        This endpoint's response contains two variables which are `style.id={layer_id}` and `style.source={source_id}`. When you use this default style in maplibre, please generate unique layer ID and soruce ID, then replace them before adding layer to maplibre.
-
-        If the style is not registered in the database yet, the endpoint will create style randomly.
-
-        **This endpoint might be restricted by the access level of the dataset**
-      security:
-        - Azure AD authentication: []
-        - API access token: []
-      parameters:
-        - schema:
-            type: string
-          in: query
-          name: colormap_name
-          description: 'Option. If specified, use this colormap to create layer style'
-    post:
-      summary: Register default layer style for a dataset
-      operationId: post-datasets-id-style-type
-      responses:
-        '200':
-          description: OK
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/DatasetLayerStyle'
-      description: |-
-        This endpoint is to register the default layer style for specified dataset.
-
-        This endpoint will replace `style.id` to `style.id={layer_id}` and `style.source` to `style.source={source_id}`. When you use this default style in maplibre, please generate unique layer ID and soruce ID, then replace them before adding layer to maplibre.
-
-        **This endpoint only can be used by the dataset owner or super user.**
-      tags:
-        - datasets
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/DatasetLayerStyle'
-            examples:
-              Example for vector:
-                value:
-                  dataset_id: 8c4810867c50ee006b11abf19876a750
-                  layer_id: drr.dynamic_subnational_hhr
-                  layer_type: fill
-                  source:
-                    type: vector
-                    url: 'https://dev.undpgeohub.org/api/vector/pgtileserv/tile.json?table=drr.dynamic_subnational_hhr&type=function'
-                    tiles:
-                      - 'https://pgtileserv.undpgeohub.org/drr.dynamic_subnational_hhr/{z}/{x}/{y}.pbf?params={"hdi_adjustment"%3A{"value"%3A5.4}%2C"working_age_pop_adjustment"%3A{"value"%3A-4.9}}'
-                  style:
-                    id: 6b45fc16-6e59-438c-ae7a-57f6dcc8b09c
-                    type: fill
-                    source: 8c4810867c50ee006b11abf19876a750
-                    source-layer: drr.dynamic_subnational_hhr
-                    minzoom: 0
-                    layout:
-                      visibility: visible
-                    paint:
-                      fill-color:
-                        type: interval
-                        property: heat_health_risk_index
-                        stops:
-                          - - 16.88
-                            - 'rgba(158, 1, 66, 1)'
-                          - - 29.78
-                            - 'rgba(249, 142, 82, 1)'
-                          - - 43.71
-                            - 'rgba(255, 255, 191, 1)'
-                          - - 58.52
-                            - 'rgba(137, 208, 165, 1)'
-                          - - 72.71
-                            - 'rgba(94, 79, 162, 1)'
-                        default: 'rgba(0,0,0,0)'
-                      fill-outline-color:
-                        property: heat_health_risk_index
-                        type: interval
-                        stops:
-                          - - 16.88
-                            - 'rgba(52, 0, 0, 1)'
-                          - - 29.78
-                            - 'rgba(108, 23, 0, 1)'
-                          - - 43.71
-                            - 'rgba(125, 127, 70, 1)'
-                          - - 58.52
-                            - 'rgba(5, 85, 50, 1)'
-                          - - 72.71
-                            - 'rgba(0, 0, 47, 1)'
-                        default: 'rgba(0,0,0,0)'
-                      fill-opacity: 1
-                  colormap_name: spectral
-                  classification_method: q
-        description: ''
-      security:
-        - Azure AD authentication: []
-        - API access token: []
-    delete:
-      summary: Delete default layer style for a dataset
-      operationId: delete-datasets-id-style-type
-      responses:
-        '204':
-          description: No Content
-      description: |-
-        This endpoint is to update the default layer style for specified dataset
-
-        **This endpoint only can be used by the dataset owner or super user.**
-      tags:
-        - datasets
-      security:
-        - Azure AD authentication: []
-        - API access token: []
   /swagger/spec.json:
     get:
       summary: Get open API JSON definition
